pyproject.toml
[tool.poetry]
name = "library"
version = "0.1.0"
description = ""
authors = ["Shitikov Kirill <shitikovkirillt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.6"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
drf-yasg = "^1.21.8"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
